AI_ML/requirements.txt
torch
torchvision
ultralytics
opencv-python
numpy
pillow
